requirements.txt
discord~=1.7.3
openai~=1.35.7
python-dotenv~=1.0.1